requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.1.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
Flask-Migrate==4.0.7
flask-session==0.8.0
flask-sqlalchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==8.2.0
importlib-resources==6.4.3
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-core==0.2.29
langchain-openai==0.1.21
langchain-text-splitters==0.2.2
langsmith==0.1.98
Mako==1.3.5
MarkupSafe==2.1.5
msgspec==0.18.6
multidict==6.0.5
numpy==1.24.4
openai==1.40.3
orjson==3.10.6
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
